requirements.txt
zeep
requests
lxml
python-dateutil
sqlalchemy
